Refractoring split types declaration #5
