+++ client/components/header/index.tsx
@@ -4,24 +4,21 @@ import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
 import Typography from "@mui/material/Typography";
 import startCase from "lodash/startCase";
+import { HeaderProps } from "@/components/header/types";
 
-interface HeaderProps {
-  title?: string;
-}
-
-const Header = ({ title = "neptune mutual" }: HeaderProps) => {
+const Header = ({
+  title = "neptune mutual",
+  imageSrc = "/images/neptune.png",
+}: HeaderProps) => {
   return (
     <Box sx={{ flexGrow: 1 }}>
       <AppBar elevation={0} position="static">
         <Toolbar>
-          <div className="mr-2 flex">
-            <Image
-              src="/images/neptune.png"
-              alt="Picture of the author"
-              width={50}
-              height={48}
-            />
-          </div>
+          {imageSrc ? (
+            <div className="mr-2 flex">
+              <Image src={imageSrc} alt="company" width={50} height={48} />
+            </div>
+          ) : null}
           <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
             {startCase(title?.toLowerCase())}
           </Typography>
