+++ client/components/header/types.ts
@@ -1,0 +1,4 @@
+export interface HeaderProps {
+  title?: string;
+  imageSrc?: string;
+}
