+++ client/pages/Home/types.ts
@@ -1,0 +1,5 @@
+export interface listTextFieldProps {
+  id: string;
+  label: string;
+  name: "primary" | "secondary";
+}
