+++ client/pages/Home/utils.ts
@@ -1,0 +1,14 @@
+import { listTextFieldProps } from "@/pages/Home/types";
+
+export const listTextField: listTextFieldProps[] = [
+  {
+    id: "outlined-primary",
+    label: "primary",
+    name: "primary",
+  },
+  {
+    id: "outlined-secondary",
+    label: "secondary",
+    name: "secondary",
+  },
+];
